Player doesn't stick to wall
I duplicated the capsule collider on the player but made it wider. Then I added a frictionless physics material to it. This way the player is frictionless when it touches the wall, however it is more wide but not as tall as the original so that it touches the wall but not the ground.